requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.1
Django==3.2.8
django-allauth==0.47.0
django-compat==1.0.15
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
idna==3.3
mysqlclient==2.1.0
numpy==1.22.0
oauthlib==3.1.1
opencv-python==4.5.5.62
Pillow==8.4.0
pipreqs==0.4.11
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
yarg==0.1.9
